pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "garmin-auth"
version = "0.1.0"
description = "Self-healing Garmin Connect OAuth authentication — login, token refresh, and rate limit recovery"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Kostas Georgiou", email = "delfinas7kostas@gmail.com" }]
keywords = ["garmin", "oauth", "authentication", "garmin-connect", "fitness"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "garminconnect>=0.2.38,<0.3.0",
    "requests>=2.31",
]

[project.optional-dependencies]
db = ["psycopg2-binary>=2.9"]
dev = ["pytest>=8.0", "pytest-mock>=3.12"]

[project.scripts]
garmin-auth = "garmin_auth.cli:main"

[project.urls]
Homepage = "https://github.com/drkostas/garmin-auth"
Repository = "https://github.com/drkostas/garmin-auth"

[tool.hatch.build.targets.wheel]
packages = ["src/garmin_auth"]
